move url processiong to Url in common package
