+++ packages/crawler/src/crawler-entry-point.spec.ts
@@ -2,35 +2,34 @@
 // Licensed under the MIT License.
 import 'reflect-metadata';
 
+import { Url } from 'common';
 import { Container } from 'inversify';
 import { GlobalLogger } from 'logger';
 import { IMock, Mock, Times } from 'typemoq';
 import { CrawlerEntryPoint } from './crawler-entry-point';
 import { CrawlerEngine } from './crawler/crawler-engine';
 import { CrawlerRunOptions } from './types/run-options';
-import { URLProcessor } from './utility/url-processor';
 
 describe(CrawlerEntryPoint, () => {
     let testSubject: CrawlerEntryPoint;
     let containerMock: IMock<Container>;
     let crawlerEngineMock: IMock<CrawlerEngine>;
     let loggerMock: IMock<GlobalLogger>;
-    let urlProcessorMock: IMock<URLProcessor>;
+    let urlMock: IMock<typeof Url>;
 
     beforeEach(() => {
         containerMock = Mock.ofType(Container);
         crawlerEngineMock = Mock.ofType(CrawlerEngine);
         loggerMock = Mock.ofType(GlobalLogger);
-        urlProcessorMock = Mock.ofType(URLProcessor);
+        urlMock = Mock.ofType<typeof Url>();
 
-        testSubject = new CrawlerEntryPoint(containerMock.object);
+        testSubject = new CrawlerEntryPoint(containerMock.object, urlMock.object);
     });
 
     it('crawl', async () => {
         const testInput: CrawlerRunOptions = { baseUrl: 'url' };
         containerMock.setup((cm) => cm.get(CrawlerEngine)).returns(() => crawlerEngineMock.object);
         containerMock.setup((c) => c.get(GlobalLogger)).returns(() => loggerMock.object);
-        containerMock.setup((c) => c.get(URLProcessor)).returns(() => urlProcessorMock.object);
         const startCommand = jest.spyOn(crawlerEngineMock.object, 'start').mockImplementationOnce(async () => Promise.resolve());
 
         loggerMock
@@ -46,7 +45,7 @@ describe(CrawlerEntryPoint, () => {
         const testInput: CrawlerRunOptions = { baseUrl: 'url?' };
         containerMock.setup((cm) => cm.get(CrawlerEngine)).returns(() => crawlerEngineMock.object);
         containerMock.setup((c) => c.get(GlobalLogger)).returns(() => loggerMock.object);
-        containerMock.setup((c) => c.get(URLProcessor)).returns(() => urlProcessorMock.object);
+        urlMock.setup((um) => um.hasQueryParameters('url?')).returns(() => true);
         const startCommand = jest.spyOn(crawlerEngineMock.object, 'start').mockImplementationOnce(async () => Promise.resolve());
 
         loggerMock
@@ -55,6 +54,6 @@ describe(CrawlerEntryPoint, () => {
             .verifiable(Times.once());
 
         await testSubject.crawl(testInput);
-        expect(startCommand).toBeCalled();
+        expect(startCommand).not.toBeCalled();
     });
 });
