+++ packages/crawler/src/crawler-entry-point.ts
@@ -1,14 +1,14 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+import { Url } from 'common';
 import { Container } from 'inversify';
 import { GlobalLogger } from 'logger';
 import { CrawlerEngine } from './crawler/crawler-engine';
 import { CrawlerRunOptions } from './types/run-options';
-import { URLProcessor } from './utility/url-processor';
 
 export class CrawlerEntryPoint {
-    constructor(private readonly container: Container) {}
+    constructor(private readonly container: Container, private readonly urlObj: typeof Url = Url) {}
 
     public async crawl(crawlerRunOptions: CrawlerRunOptions): Promise<void> {
         const logger = this.container.get(GlobalLogger);
@@ -24,8 +24,6 @@ export class CrawlerEntryPoint {
     }
 
     private isBaseUrlValid(baseUrl: string): boolean {
-        const urlProcessor = this.container.get(URLProcessor);
-
-        return urlProcessor.hasQueryParameters(baseUrl);
+        return this.urlObj.hasQueryParameters(baseUrl);
     }
 }
