+++ packages/crawler/src/setup-crawler-container.spec.ts
@@ -15,7 +15,6 @@ import { PageProcessorFactory } from './page-processors/page-processor-factory';
 import { PageScanner } from './scanner-operations/page-scanner';
 import { setupCrawlerContainer } from './setup-crawler-container';
 import { ActiveElementsFinder } from './utility/active-elements-finder';
-import { URLProcessor } from './utility/url-processor';
 
 describe(setupCrawlerContainer, () => {
     it('resolves dependencies', () => {
@@ -33,6 +32,5 @@ describe(setupCrawlerContainer, () => {
         expect(container.get(AxePuppeteerFactory)).toBeDefined();
         expect(container.get(ActiveElementsFinder)).toBeDefined();
         expect(container.get(GlobalLogger)).toBeDefined();
-        expect(container.get(URLProcessor)).toBeDefined();
     });
 });
