+++ packages/crawler/src/setup-crawler-container.ts
@@ -10,13 +10,11 @@ import { AxePuppeteerFactory } from './axe-puppeteer/axe-puppeteer-factory';
 import { CrawlerConfiguration } from './crawler/crawler-configuration';
 import { CrawlerFactory } from './crawler/crawler-factory';
 import { PageProcessorFactory } from './page-processors/page-processor-factory';
-import { URLProcessor } from './utility/url-processor';
 
 export function setupCrawlerContainer(): inversify.Container {
     const container = new inversify.Container({ autoBindInjectable: true });
     container.bind(CrawlerConfiguration).toSelf().inSingletonScope();
     container.bind(CrawlerFactory).toSelf().inSingletonScope();
-    container.bind(URLProcessor).toSelf().inSingletonScope();
     container.bind(ApifyResourceCreator).toSelf().inSingletonScope();
     container.bind(PageProcessorFactory).toSelf().inSingletonScope();
     container.bind('ReporterFactory').toConstantValue(reporterFactory);
